refactor: Integrate jwt and auth
