+++ src/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { JwtPayload } from '../jwt/jwt.strategy';
+import { JwtPayload } from './strategies/jwt.strategy';
 
 @Injectable()
 export class AuthService {
@@ -10,7 +10,10 @@ export class AuthService {
     const payload: JwtPayload = { email, id };
 
     // TODO: Using refresh-token
-    const accessToken = this.jwtService.sign(payload);
+    const accessToken = this.jwtService.sign(payload, {
+      expiresIn: '2h',
+      secret: process.env.JWT_SECRET,
+    });
 
     return accessToken;
   }
