--- src/jwt/jwt.module.ts
@@ -1,20 +1,0 @@
-import { Module } from '@nestjs/common';
-import { JwtModule } from '@nestjs/jwt';
-import { AuthService } from '../auth/auth.service';
-import { JwtStrategy } from './jwt.strategy';
-
-@Module({
-  imports: [
-    JwtModule.registerAsync({
-      useFactory: () => ({
-        secret: process.env.JWT_SECRET,
-        signOptions: {
-          expiresIn: process.env.JWT_EXPIRES_IN,
-        },
-      }),
-    }),
-  ],
-  providers: [AuthService, JwtStrategy],
-  exports: [JwtModule, AuthService],
-})
-export class JwtAuthModule {}
